Move implementations into data_manager
