+++ src/analyses/histgrowth.rs
@@ -4,8 +4,8 @@ use std::io::Write;
 use clap::{arg, value_parser, Arg, Command};
 use rayon::iter::{ParallelBridge, ParallelIterator};
 
-use crate::{analyses::InputRequirement, hist::HistAuxilliary, io::{write_histgrowth_table, write_table}, util::CountType};
-use crate::{abacus::ViewParams, clap_enum_variants, io::OutputFormat};
+use crate::{analyses::InputRequirement, data_manager::{HistAuxilliary, ViewParams}, io::write_table, util::CountType};
+use crate::{clap_enum_variants, io::OutputFormat};
 
 use super::{Analysis, ReportSection};
