+++ src/cli.rs
@@ -4,9 +4,7 @@ use std::io::{Error, ErrorKind};
 use std::str::FromStr;
 
 /* external crate */
-use clap::{crate_version, Command, Parser, Subcommand};
-use rayon::prelude::*;
-use strum::VariantNames;
+use clap::{crate_version, ArgMatches, Command};
 
 use crate::analyses::growth::Growth;
 use crate::analyses::histgrowth::Histgrowth;
@@ -16,7 +14,6 @@ use crate::analyses::ordered_histgrowth::OrderedHistgrowth;
 use crate::analyses::table::Table;
 use crate::analyses::{self, Analysis};
 use crate::data_manager::DataManager;
-use crate::io::*;
 use crate::util::*;
 
 pub enum RequireThreshold {
@@ -44,452 +41,6 @@ macro_rules! clap_enum_variants_no_all {
         clap::builder::PossibleValuesParser::new(<$e>::VARIANTS.iter().filter(|&x| x != &"all"))
             .map(|s| s.parse::<$e>().unwrap())
     }};
-}
-
-#[derive(Subcommand, Debug)]
-pub enum Params {
-    #[clap(alias = "I", about = "Return general graph and paths info")]
-    Info {
-        #[clap(
-            index = 1,
-            help = "graph in GFA1 format, accepts also compressed (.gz) file",
-            required = true
-        )]
-        gfa_file: String,
-        #[clap(
-            name = "subset",
-            short,
-            long,
-            help = "Produce counts by subsetting the graph to a given list of paths (1-column list) or path coordinates (3- or 12-column BED file)",
-            default_value = ""
-        )]
-        positive_list: String,
-        #[clap(
-            name = "exclude",
-            short,
-            long,
-            help = "Exclude bp/node/edge in growth count that intersect with paths (1-column list) or path coordinates (3- or 12-column BED-file) provided by the given file; all intersecting bp/node/edge will be exluded also in other paths not part of the given list",
-            default_value = ""
-        )]
-        negative_list: String,
-        #[clap(
-            short,
-            long,
-            help = "Merge counts from paths by path-group mapping from given tab-separated two-column file",
-            default_value = ""
-        )]
-        groupby: String,
-        #[clap(
-            short = 'H',
-            long,
-            help = "Merge counts from paths belonging to same haplotype"
-        )]
-        groupby_haplotype: bool,
-        #[clap(
-            short = 'S',
-            long,
-            help = "Merge counts from paths belonging to same sample"
-        )]
-        groupby_sample: bool,
-        #[clap(short, long, help = "Choose output format: table (tab-separated-values) or html report", 
-            default_value = "table",
-            ignore_case = true,
-            value_parser = clap_enum_variants!(OutputFormat),)]
-        output_format: OutputFormat,
-        #[clap(
-            short,
-            long,
-            help = "Run in parallel on N threads (0 for number of CPU cores)",
-            default_value = "0"
-        )]
-        threads: usize,
-    },
-
-    #[clap(alias = "hg", about = "Run hist and growth. Return the growth curve")]
-    Histgrowth {
-        #[clap(
-            index = 1,
-            help = "graph in GFA1 format, accepts also compressed (.gz) file",
-            required = true
-        )]
-        gfa_file: String,
-        #[clap(short, long, help = "Graph quantity to be counted", default_value = "node", ignore_case = true, value_parser = clap_enum_variants!(CountType),)]
-        count: CountType,
-        #[clap(
-            name = "subset",
-            short,
-            long,
-            help = "Produce counts by subsetting the graph to a given list of paths (1-column list) or path coordinates (3- or 12-column BED file)",
-            default_value = ""
-        )]
-        positive_list: String,
-        #[clap(
-            name = "exclude",
-            short,
-            long,
-            help = "Exclude bp/node/edge in growth count that intersect with paths (1-column list) or path coordinates (3- or 12-column BED-file) provided by the given file",
-            default_value = ""
-        )]
-        negative_list: String,
-        #[clap(
-            short,
-            long,
-            help = "Merge counts from paths by path-group mapping from given tab-separated two-column file",
-            default_value = ""
-        )]
-        groupby: String,
-        #[clap(
-            short = 'H',
-            long,
-            help = "Merge counts from paths belonging to same haplotype"
-        )]
-        groupby_haplotype: bool,
-        #[clap(
-            short = 'S',
-            long,
-            help = "Merge counts from paths belonging to same sample"
-        )]
-        groupby_sample: bool,
-        #[clap(
-            short = 'l',
-            long,
-            help = "Ignore all countables with a coverage lower than the specified threshold. The coverage of a countable corresponds to the number of path/walk that contain it. Repeated appearances of a countable in the same path/walk are counted as one. You can pass a comma-separated list of coverage thresholds, each one will produce a separated growth curve (e.g., --coverage 2,3). Use --quorum to set a threshold in conjunction with each coverage (e.g., --quorum 0.5,0.9)",
-            default_value = "1"
-        )]
-        coverage: String,
-        #[clap(
-            short,
-            long,
-            help = "Unlike the --coverage parameter, which specifies a minimum constant number of paths for all growth point m (1 <= m <= num_paths), --quorum adjust the threshold based on m. At each m, a countable is counted in the average growth if the countable is contained in at least floor(m*quorum) paths. Example: A quorum of 0.9 requires a countable to be in 90% of paths for each subset size m. At m=10, it must appear in at least 9 paths. At m=100, it must appear in at least 90 paths. A quorum of 1 (100%) requires presence in all paths of the subset, corresponding to the core. Default: 0, a countable counts if it is present in any path at each growth point. Specify multiple quorum values with a comma-separated list (e.g., --quorum 0.5,0.9). Use --coverage to set static path thresholds in conjunction with variable quorum percentages (e.g., --coverage 5,10).",
-            default_value = "0"
-        )]
-        quorum: String,
-        #[clap(short = 'a', long, help = "Also include histogram in output")]
-        hist: bool,
-        #[clap(short, long, help = "Choose output format: table (tab-separated-values) or html report", default_value = "table", ignore_case = true, value_parser = clap_enum_variants!(OutputFormat),)]
-        output_format: OutputFormat,
-        #[clap(
-            short,
-            long,
-            help = "Run in parallel on N threads (0 for number of CPU cores)",
-            default_value = "0"
-        )]
-        threads: usize,
-    },
-
-    #[clap(alias = "h", about = "Calculate coverage histogram")]
-    Hist {
-        #[clap(
-            index = 1,
-            help = "graph in GFA1 format, accepts also compressed (.gz) file",
-            required = true
-        )]
-        gfa_file: String,
-        #[clap(short, long, help = "Graph quantity to be counted", default_value = "node", ignore_case = true, value_parser = clap_enum_variants!(CountType),)]
-        count: CountType,
-        #[clap(
-            name = "subset",
-            short,
-            long,
-            help = "Produce counts by subsetting the graph to a given list of paths (1-column list) or path coordinates (3- or 12-column BED file)",
-            default_value = ""
-        )]
-        positive_list: String,
-        #[clap(
-            name = "exclude",
-            short,
-            long,
-            help = "Exclude bp/node/edge in growth count that intersect with paths (1-column list) or path coordinates (3- or 12-column BED-file) provided by the given file; all intersecting bp/node/edge will be exluded also in other paths not part of the given list",
-            default_value = ""
-        )]
-        negative_list: String,
-        #[clap(
-            short,
-            long,
-            help = "Merge counts from paths by path-group mapping from given tab-separated two-column file",
-            default_value = ""
-        )]
-        groupby: String,
-        #[clap(
-            short = 'H',
-            long,
-            help = "Merge counts from paths belonging to same haplotype"
-        )]
-        groupby_haplotype: bool,
-        #[clap(
-            short = 'S',
-            long,
-            help = "Merge counts from paths belonging to same sample"
-        )]
-        groupby_sample: bool,
-        #[clap(short, long, help = "Choose output format: table (tab-separated-values) or html report", default_value = "table", ignore_case = true, value_parser = clap_enum_variants!(OutputFormat),)]
-        output_format: OutputFormat,
-        #[clap(
-            short,
-            long,
-            help = "Run in parallel on N threads (0 for number of CPU cores)",
-            default_value = "0"
-        )]
-        threads: usize,
-    },
-
-    #[clap(alias = "g", about = "Calculate growth curve from coverage histogram")]
-    Growth {
-        #[clap(
-            index = 1,
-            help = "Coverage histogram as tab-separated value (tsv) file",
-            required = true
-        )]
-        hist_file: String,
-        #[clap(
-            short = 'l',
-            long,
-            help = "Ignore all countables with a coverage lower than the specified threshold. The coverage of a countable corresponds to the number of path/walk that contain it. Repeated appearances of a countable in the same path/walk are counted as one. You can pass a comma-separated list of coverage thresholds, each one will produce a separated growth curve (e.g., --coverage 2,3). Use --quorum to set a threshold in conjunction with each coverage (e.g., --quorum 0.5,0.9)",
-            default_value = "1"
-        )]
-        coverage: String,
-        #[clap(
-            short,
-            long,
-            help = "Unlike the --coverage parameter, which specifies a minimum constant number of paths for all growth point m (1 <= m <= num_paths), --quorum adjust the threshold based on m. At each m, a countable is counted in the average growth if the countable is contained in at least floor(m*quorum) paths. Example: A quorum of 0.9 requires a countable to be in 90% of paths for each subset size m. At m=10, it must appear in at least 9 paths. At m=100, it must appear in at least 90 paths. A quorum of 1 (100%) requires presence in all paths of the subset, corresponding to the core. Default: 0, a countable counts if it is present in any path at each growth point. Specify multiple quorum values with a comma-separated list (e.g., --quorum 0.5,0.9). Use --coverage to set static path thresholds in conjunction with variable quorum percentages (e.g., --coverage 5,10).",
-            default_value = "0"
-        )]
-        quorum: String,
-        #[clap(short = 'a', long, help = "Also include histogram in output")]
-        hist: bool,
-        #[clap(short, long, help = "Choose output format: table (tab-separated-values) or html report", default_value = "table", ignore_case = true, value_parser = clap_enum_variants!(OutputFormat),)]
-        output_format: OutputFormat,
-        #[clap(
-            short,
-            long,
-            help = "Run in parallel on N threads (0 for number of CPU cores)",
-            default_value = "0"
-        )]
-        threads: usize,
-    },
-
-    #[clap(
-        alias = "o",
-        about = "Calculate growth curve based on group file order (if order is unspecified, use path order in GFA)"
-    )]
-    OrderedHistgrowth {
-        #[clap(
-            index = 1,
-            help = "graph in GFA1 format, accepts also compressed (.gz) file",
-            required = true
-        )]
-        gfa_file: String,
-        #[clap(short, long, help = "Graph quantity to be counted", default_value = "node", ignore_case = true, value_parser = clap_enum_variants_no_all!(CountType),)]
-        count: CountType,
-        #[clap(
-            name = "order",
-            short = 'O',
-            long,
-            help = "The ordered histogram will be produced according to order of paths/groups in the supplied file (1-column list). If this option is not used, the order is determined by the rank of paths/groups in the subset list, and if that option is not used, the order is determined by the rank of paths/groups in the GFA file.",
-            default_value = ""
-        )]
-        order: String,
-        #[clap(
-            name = "subset",
-            short,
-            long,
-            help = "Produce counts by subsetting the graph to a given list of paths (1-column list) or path coordinates (3- or 12-column BED file). If the \"order\" option is not used, the subset list will also indicate the order of paths/groups in the histogram.",
-            default_value = ""
-        )]
-        positive_list: String,
-        #[clap(
-            name = "exclude",
-            short,
-            long,
-            help = "Exclude bp/node/edge in growth count that intersect with paths (1-column list) or path coordinates (3- or 12-column BED-file) provided by the given file",
-            default_value = ""
-        )]
-        negative_list: String,
-        #[clap(
-            short,
-            long,
-            help = "Merge counts from paths by path-group mapping from given tab-separated two-column file",
-            default_value = ""
-        )]
-        groupby: String,
-        #[clap(
-            short = 'H',
-            long,
-            help = "Merge counts from paths belonging to same haplotype"
-        )]
-        groupby_haplotype: bool,
-        #[clap(
-            short = 'S',
-            long,
-            help = "Merge counts from paths belonging to same sample"
-        )]
-        groupby_sample: bool,
-        #[clap(
-            short,
-            long,
-            help = "List of quorum fractions of the form <level1>,<level2>,... Number of values must be one or match that of coverage setting",
-            default_value = "0"
-        )]
-        quorum: String,
-        #[clap(
-            short = 'l',
-            long,
-            help = "List of absolute coverage thresholds of the form <level1>,<level2>,... Number of values must be one or match that of quorum setting",
-            default_value = "1"
-        )]
-        coverage: String,
-        #[clap(short, long, help = "Choose output format: table (tab-separated-values) or html report", default_value = "table", ignore_case = true, value_parser = clap_enum_variants!(OutputFormat),)]
-        output_format: OutputFormat,
-        #[clap(
-            short,
-            long,
-            help = "Run in parallel on N threads (0 for number of CPU cores)",
-            default_value = "0"
-        )]
-        threads: usize,
-    },
-
-    #[clap(about = "Compute coverage table for count type")]
-    Table {
-        #[clap(
-            index = 1,
-            help = "graph in GFA1 format, accepts also compressed (.gz) file",
-            required = true
-        )]
-        gfa_file: String,
-        #[clap(short, long, help = "Graph quantity to be counted", default_value = "node", ignore_case = true, value_parser = clap_enum_variants_no_all!(CountType),)]
-        count: CountType,
-        #[clap(
-            name = "total",
-            short = 'a',
-            long,
-            help = "Summarize by totaling presence/absence over all groups"
-        )]
-        total: bool,
-        #[clap(
-            name = "subset",
-            short,
-            long,
-            help = "Produce counts by subsetting the graph to a given list of paths (1-column list) or path coordinates (3- or 12-column BED file)",
-            default_value = ""
-        )]
-        positive_list: String,
-        #[clap(
-            name = "exclude",
-            short,
-            long,
-            help = "Exclude bp/node/edge in growth count that intersect with paths (1-column list) or path coordinates (3- or 12-column BED-file) provided by the given file",
-            default_value = ""
-        )]
-        negative_list: String,
-        #[clap(
-            short,
-            long,
-            help = "Merge counts from paths by path-group mapping from given tab-separated two-column file",
-            default_value = ""
-        )]
-        groupby: String,
-        #[clap(
-            short = 'H',
-            long,
-            help = "Merge counts from paths belonging to same haplotype"
-        )]
-        groupby_haplotype: bool,
-        #[clap(
-            short = 'S',
-            long,
-            help = "Merge counts from paths belonging to same sample"
-        )]
-        groupby_sample: bool,
-        #[clap(
-            short,
-            long,
-            help = "Run in parallel on N threads (0 for number of CPU cores)",
-            default_value = "0"
-        )]
-        threads: usize,
-    },
-    //#[clap(
-    //    alias = "C",
-    //    about = "Calculate the histogram and growth of a Compacted de Bruijn Graph"
-    //)]
-    //Cdbg {
-    //    #[clap(
-    //        index = 1,
-    //        help = "graph in GFA1 format, accepts also compressed (.gz) file representing a compacted de Bruijn graph",
-    //        required = true
-    //    )]
-    //    gfa_file: String,
-    //    #[clap(short, long, help = "Value of k for cdBG", default_value = "")]
-    //    k: usize,
-    //    #[clap(
-    //        name = "subset",
-    //        short,
-    //        long,
-    //        help = "Produce counts by subsetting the graph to a given list of paths (1-column list) or path coordinates (3- or 12-column BED file)",
-    //        default_value = ""
-    //    )]
-    //    positive_list: String,
-    //    #[clap(
-    //        name = "exclude",
-    //        short,
-    //        long,
-    //        help = "Exclude bp/node/edge in growth count that intersect with paths (1-column list) or path coordinates (3- or 12-column BED-file) provided by the given file",
-    //        default_value = ""
-    //    )]
-    //    negative_list: String,
-    //    #[clap(
-    //        short,
-    //        long,
-    //        help = "Merge counts from paths by path-group mapping from given tab-separated two-column file",
-    //        default_value = ""
-    //    )]
-    //    #[clap(
-    //        short,
-    //        long,
-    //        help = "Merge counts from paths by path-group mapping from given tab-separated two-column file",
-    //        default_value = ""
-    //    )]
-    //    groupby: String,
-    //    #[clap(
-    //        short = 'H',
-    //        long,
-    //        help = "Merge counts from paths belonging to same haplotype"
-    //    )]
-    //    groupby_haplotype: bool,
-    //    #[clap(
-    //        short = 'S',
-    //        long,
-    //        help = "Merge counts from paths belonging to same sample"
-    //    )]
-    //    groupby_sample: bool,
-    //    #[clap(
-    //        short,
-    //        long,
-    //        help = "Run in parallel on N threads (0 for number of CPU cores)",
-    //        default_value = "0"
-    //    )]
-    //    threads: usize,
-    //},
-}
-
-impl Params {
-    pub fn test_default_histgrowth() -> Self {
-        Params::Histgrowth {
-            gfa_file: String::new(),
-            count: CountType::Node,
-            positive_list: String::new(),
-            negative_list: String::new(),
-            groupby: String::new(),
-            groupby_haplotype: false,
-            groupby_sample: false,
-            coverage: "1".to_string(),
-            quorum: "0".to_string(),
-            hist: false,
-            output_format: OutputFormat::Table,
-            threads: 0,
-        }
-    }
 }
 
 pub fn parse_threshold_cli(
@@ -550,40 +101,17 @@ pub fn parse_threshold_cli(
 // https://github.com/rayon-rs/rayon/issues/878
 // We run this function in the main otherwise in the tests the second time we run the function
 // "run" it will crush
-pub fn set_number_of_threads(params: &Params) {
-    if let Params::Histgrowth { threads, .. }
-    | Params::Hist { threads, .. }
-    | Params::Info { threads, .. }
-    | Params::OrderedHistgrowth { threads, .. }
-    | Params::Table { threads, .. }
-    //| Params::Cdbg { threads, .. } 
-    = params {
+pub fn set_number_of_threads(params: &ArgMatches) {
         //if num_threads is 0 then the Rayon will select 
         //the number of threads to the core number automatically 
+        let subcommand_name = params.subcommand_name().unwrap();
+        let matches = params.subcommand_matches(subcommand_name).unwrap();
+        let threads = matches.get_one("threads").unwrap();
         rayon::ThreadPoolBuilder::new()
             .num_threads(*threads)
             .build_global()
             .expect("Failed to initialize global thread pool");
         log::info!("running panacus on {} threads", rayon::current_num_threads());
-    }
-}
-
-// make sure either group, groupby-sample, or groupby-haplotype is set
-pub fn validate_single_groupby_option(
-    groupby: &str,
-    groupby_haplotype: bool,
-    groupby_sample: bool,
-) -> Result<(), Error> {
-    let mut c = 0;
-    c += (!groupby.is_empty()) as u8;
-    c += (groupby_haplotype) as u8;
-    c += (groupby_sample) as u8;
-    if c > 1 {
-        let msg = "At most one option of groupby, groupby-haplotype, and groupby-sample can be set at once, but at least two are given.";
-        log::error!("{}", &msg);
-        return Err(Error::new(ErrorKind::InvalidInput, msg));
-    }
-    Ok(())
 }
 
 pub fn run<W: Write>(out: &mut BufWriter<W>) -> Result<(), Error> {
@@ -599,6 +127,7 @@ pub fn run<W: Write>(out: &mut BufWriter<W>) -> Result<(), Error> {
         .subcommand(Growth::get_subcommand())
         .get_matches();
 
+    set_number_of_threads(&matches);
     if let Some((req, view_params, gfa_file)) = Info::get_input_requirements(&matches) {
         let dm = DataManager::from_gfa_with_view(&gfa_file, req, &view_params)?;
         let mut info = Info::build(&dm, &matches)?;
@@ -904,43 +433,5 @@ mod tests {
         let result = parse_threshold_cli(threshold_str, RequireThreshold::Relative);
         assert!(result.is_err());
         assert_eq!(result.unwrap_err().kind(), ErrorKind::InvalidData);
-    }
-
-    #[test]
-    fn test_validate_single_groupby_option() {
-        let test_cases = vec![
-            // Valid cases
-            ("", false, false, true),       // None set
-            ("group1", false, false, true), // Only groupby is set
-            ("", true, false, true),        // Only groupby_haplotype is set
-            ("", false, true, true),        // Only groupby_sample is set
-            // Invalid cases
-            ("group1", true, false, false), // groupby and groupby_haplotype set
-            ("group1", false, true, false), // groupby and groupby_sample set
-            ("", true, true, false),        // groupby_haplotype and groupby_sample set
-            ("group1", true, true, false),  // All options set
-        ];
-
-        for (test_groupby, test_groupby_haplotype, test_groupby_sample, should_pass) in test_cases {
-            //let mut params = Params::test_default_histgrowth();
-            let result = validate_single_groupby_option(
-                test_groupby,
-                test_groupby_haplotype,
-                test_groupby_sample,
-            );
-            if should_pass {
-                assert!(
-                    result.is_ok(),
-                    "Expected OK, but got error for input: groupby = '{}', groupby_haplotype = {}, groupby_sample = {}",
-                    test_groupby, test_groupby_haplotype, test_groupby_sample
-                );
-            } else {
-                assert!(
-                    result.is_err(),
-                    "Expected error, but got OK for input: groupby = '{}', groupby_haplotype = {}, groupby_sample = {}",
-                    test_groupby, test_groupby_haplotype, test_groupby_sample
-                );
-            }
-        }
     }
 }
